Enhance Dump

+++ frameworks/native/audioeffect/include/audio_enhance_chain.h
@@ -102,7 +102,7 @@ private:
     void InitAudioEnhanceChain();
     void InitDump();
     void ReleaseEnhanceChain();
-    void WriteDumpFile(std::unique_ptr<EnhanceBuffer> &enhanceBuffer, uint64_t length);
+    void WriteDumpFile(std::unique_ptr<EnhanceBuffer> &enhanceBuffer, uint32_t length);
     int32_t GetOneFrameInputData(std::unique_ptr<EnhanceBuffer> &enhanceBuffer);
     int32_t DeinterleaverData(uint8_t *src, uint32_t channel, uint8_t *dst, uint32_t offset);
     int32_t SetPropertyToHandle(AudioEffectHandle handle, const std::string &property);
